refactor(site): move images and fonts to site folder
