+++ src/site/components/Frame/Frame.module.css
@@ -4,21 +4,21 @@
 
 .thin {
   border: 2px solid var(--color-text);
-  border-image-source: url(../../../../assets/images/frame-thin.png);
+  border-image-source: url(../../images/frame-thin.png);
   border-image-slice: 2;
   border-image-repeat: repeat;
 }
 
 .thick {
   border: 4px solid var(--color-text);
-  border-image-source: url(../../../../assets/images/frame-thick.png);
+  border-image-source: url(../../images/frame-thick.png);
   border-image-slice: 4;
   border-image-repeat: repeat;
 }
 
 .button {
   border: 4px solid var(--color-text);
-  border-image-source: url(../../../../assets/images/frame-button.png);
+  border-image-source: url(../../images/frame-button.png);
   border-image-slice: 4;
   border-image-repeat: repeat;
 }
