+++ src/site/components/GoldIcon/GoldIcon.module.css
@@ -3,6 +3,6 @@
   width: 16px;
   height: 16px;
   vertical-align: top;
-  background: url(../../../../assets/images/gold.png) no-repeat;
+  background: url(../../images/gold.png) no-repeat;
   background-size: 100% 100%;
 }
