+++ src/site/components/Page/Page.css
@@ -1,6 +1,6 @@
 @font-face {
   font-family: 'MysticCards';
-  src: url(../../../../assets/fonts/MysticCards.ttf) format('truetype');
+  src: url(../../fonts/MysticCards.ttf) format('truetype');
   font-display: swap;
 }
