+++ src/site/images/gold.png
