+++ src/site/images/icon-combat.png
