+++ src/site/images/icon-fatigue.png
